% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{aggregate_to_district}
\alias{aggregate_to_district}
\title{Aggregate campus data to district level}
\usage{
aggregate_to_district(campus_df, end_year)
}
\arguments{
\item{campus_df}{Campus-level data frame}

\item{end_year}{School year end}
}
\value{
District-level data frame
}
\description{
Creates district-level summaries from campus-level data.
}
\keyword{internal}
