% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{build_assessment_url}
\alias{build_assessment_url}
\title{Build assessment URL}
\usage{
build_assessment_url(end_year, subject, grade, level)
}
\arguments{
\item{end_year}{School year end}

\item{subject}{"ela" or "math"}

\item{grade}{Grade level}

\item{level}{"district" or "school"}
}
\value{
URL string
}
\description{
Constructs the URL for downloading assessment data files based on year and pattern.
}
\keyword{internal}
