% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{build_fast_url}
\alias{build_fast_url}
\title{Build FAST era URL (2023-2025)}
\usage{
build_fast_url(end_year, subject, grade, level, grade_str, year_suffix)
}
\description{
Build FAST era URL (2023-2025)
}
\keyword{internal}
