% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{build_fsa_url}
\alias{build_fsa_url}
\title{Build FSA era URL (2019, 2022)}
\usage{
build_fsa_url(end_year, subject, grade, level, grade_str, year_suffix)
}
\description{
Build FSA era URL (2019, 2022)
}
\keyword{internal}
