% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_fte_data}
\alias{download_fte_data}
\title{Download FTE enrollment data (2008-2013)}
\usage{
download_fte_data(end_year)
}
\arguments{
\item{end_year}{School year end (2008-2013)}
}
\value{
List with campus and district data frames
}
\description{
Downloads FTE (Full-Time Equivalent) enrollment files from FLDOE.
These files have district-level data with less demographic detail.
}
\keyword{internal}
