% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\docType{data}
\name{FLDOE_GRAD_BASE_URL}
\alias{FLDOE_GRAD_BASE_URL}
\title{Base URL for FLDOE graduation rate files}
\format{
An object of class \code{character} of length 1.
}
\usage{
FLDOE_GRAD_BASE_URL
}
\description{
Base URL for FLDOE graduation rate files
}
\keyword{internal}
