% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_fl_school_id}
\alias{format_fl_school_id}
\title{Format Florida School ID}
\usage{
format_fl_school_id(district_num, school_num)
}
\arguments{
\item{district_num}{District number (2 digits)}

\item{school_num}{School number (4 digits)}
}
\value{
Character string in format "DD-SSSS"
}
\description{
Creates a standardized school ID from district and school numbers.
}
\keyword{internal}
