% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_graduation.R
\name{get_available_grad_years}
\alias{get_available_grad_years}
\title{Get available graduation years}
\usage{
get_available_grad_years()
}
\value{
Integer vector of years (2016, 2017, ..., 2024)
}
\description{
Returns a vector of years for which graduation rate data is available
from the Florida Department of Education.
}
\examples{
\dontrun{
get_available_grad_years()
# Returns: 2016 2017 2018 2019 2020 2021 2022 2023 2024
}
}
