% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_cache_path}
\alias{get_cache_path}
\title{Get cache file path for given key}
\usage{
get_cache_path(key)
}
\arguments{
\item{key}{Cache key (e.g., "enr_tidy_2024" or "graduation_2024")}
}
\value{
Full path to cache file
}
\description{
Get cache file path for given key
}
\keyword{internal}
