% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_fldoe_url}
\alias{get_fldoe_url}
\title{Get URL for FLDOE membership file}
\usage{
get_fldoe_url(end_year, file_type = "membership")
}
\arguments{
\item{end_year}{School year end}

\item{file_type}{Type of file: "membership", "grade", or "race"}
}
\value{
URL string
}
\description{
Constructs the URL for downloading membership data files.
}
\keyword{internal}
