% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from FLDOE}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
List with campus and district data frames
}
\description{
Downloads school enrollment data from Florida Department of Education.
Uses membership files for 2014+ and FTE files for 2008-2013.
}
\keyword{internal}
