% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{id_assessment_aggs}
\alias{id_assessment_aggs}
\title{Identify assessment aggregation levels}
\usage{
id_assessment_aggs(df)
}
\arguments{
\item{df}{Assessment data frame}
}
\value{
Data frame with aggregation flags
}
\description{
Adds aggregation flags to assessment data.
}
