% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_graduation.R
\name{process_sheet}
\alias{process_sheet}
\title{Process a single sheet (district or school)}
\usage{
process_sheet(sheet_data, level, end_year)
}
\arguments{
\item{sheet_data}{Data frame from Excel sheet}

\item{level}{Either "District" or "School"}

\item{end_year}{School year end}
}
\value{
Data frame with processed data
}
\description{
Process a single sheet (district or school)
}
\keyword{internal}
