% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{standardize_assessment_colnames}
\alias{standardize_assessment_colnames}
\title{Standardize assessment column names}
\usage{
standardize_assessment_colnames(colnames)
}
\arguments{
\item{colnames}{Character vector of column names}
}
\value{
Character vector of standardized column names
}
\description{
Converts FLDOE assessment column names to a consistent standard.
}
\keyword{internal}
